compilation fixes of some of base dependend apps

+++ libSzarp2/include/sz4/rpn_param_entry.h
@@ -79,7 +79,7 @@ public:
 			empty_fetch_functor empty_fetch;
 
 			double v = szb_definable_calculate(stack,
-					sizeof(stack) / sizeof(stack[0]),
+					int(sizeof(stack) / sizeof(stack[0])),
 					varrarr,
 					f_cache,
 					formula,
